fix(libfdt): typecast operands to match data type
This corrects the MISRA violation C2012-10.3:
The value of an expression shall not be assigned to an object with a
narrower essential type or of a different essential type category.
The operand is explicitly typecasted to the required type
to match the operand/return type.

State: waiting
Link: https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/42598

+++ common/fdt_fixup.c
@@ -32,7 +32,7 @@
 
 static int append_psci_compatible(void *fdt, int offs, const char *str)
 {
-	return fdt_appendprop(fdt, offs, "compatible", str, strlen(str) + 1);
+	return fdt_appendprop(fdt, offs, "compatible", str, (int)strlen(str) + 1);
 }
 
 /*
